+++ Model/HardcodedSimulator/ScheduleListExact.java
@@ -13,8 +13,8 @@ public class ScheduleListExact {
     public POI[] pOIs;
     public short[] fromHomeFreqs;
     public short[] fromWorkFreqs;
-    public short[] fromHomeFreqsCDF;
-    public short[] fromWorkFreqsCDF;
+    public float[] fromHomeFreqsCDF;
+    public float[] fromWorkFreqsCDF;
     public float sumHomeFreqs=0;
     public float sumWorkFreqs=0;
 }
